requirements.txt
joblib==1.5.1
numpy==2.2.6
paho-mqtt==2.1.0
pandas==2.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
